+++ client/src/shared/apis/client.ts
@@ -1,14 +1,10 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import axios, { AxiosRequestConfig, isAxiosError } from 'axios';
-
-import Router from 'next/router';
+import axios from 'axios';
 
 import Cookies from 'js-cookie';
 
-import { ERROR_MESSAGES, TOKEN_KEYS } from '@/src/shared/const';
+import { TOKEN_KEYS } from '@/src/shared/const';
 
-import { tokenReissue } from './tokenReissue';
-
+import { errorManager, serverDisconnected } from './errorhandlers';
 import { SERVER } from '../config/apis';
 
 export const API = axios.create({
@@ -23,6 +19,7 @@ API.interceptors.request.use(
     if (Cookies.get(TOKEN_KEYS.ACCESS_TOKEN)) {
       const newConfig = { ...config };
       newConfig.headers.Authorization = Cookies.get(TOKEN_KEYS.ACCESS_TOKEN);
+      return newConfig;
     }
     return config;
   },
@@ -40,56 +37,8 @@ API.interceptors.response.use(
     const { status, config } = originRequest;
     const errorMessage = originRequest.data.message;
     if (status === 401) {
-      handleErrorMessage(errorMessage, config);
+      errorManager(errorMessage, config);
     }
     return Promise.reject(error);
   },
 );
-
-const handleErrorMessage = (
-  message: string,
-  originRequest: AxiosRequestConfig,
-) => {
-  message === ERROR_MESSAGES.JWT_EXPIRED && jwtExpired(originRequest);
-};
-
-const unauthorized = () => {
-  Router.replace('/login');
-};
-
-const serverDisconnected = (message: string) =>
-  (message === 'Network Error'
-    ? new Error('서버와 연결이 끊겼습니다. 잠시후 다시시도해주세요.')
-    : new Error('원인 파악이 불명한 에러가 발생했습니다.'));
-
-const jwtExpired = async (originRequest: AxiosRequestConfig) => {
-  try {
-    const response = await tokenReissue();
-    if (response) {
-      const { access } = response.data;
-      Cookies.set(TOKEN_KEYS.ACCESS_TOKEN, access);
-      await reRequest(originRequest);
-    }
-  } catch (error) {
-    if (isAxiosError(error)) {
-      const errorMessage = error.response?.data?.message;
-      errorMessage && unauthorized();
-    }
-  }
-};
-
-export const reRequest = async (
-  originRequest: AxiosRequestConfig & { retryFetch?: number },
-) => {
-  const requestWithRetry = {
-    ...originRequest,
-    retryFetch: originRequest.retryFetch || 0,
-  };
-
-  if (requestWithRetry.retryFetch >= 3) {
-    throw new Error('요청 제한 횟수 초과');
-  }
-
-  requestWithRetry.retryFetch += 1;
-  return API.request(requestWithRetry);
-};
