+++ client/src/shared/apis/errorhandlers/index.ts
@@ -1,0 +1,2 @@
+export { errorManager } from './errorManager';
+export { serverDisconnected } from './serverDisconnected';
